Split code into modules.
