+++ app/terminationhandler/terminationhandler.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+
